refactor: create internal/cli module
