+++ internal/features/features_test.go
@@ -82,117 +82,6 @@ func TestEnabled(t *testing.T) {
 	}
 }
 
-func TestCommandAdder(t *testing.T) {
-	tests := []struct {
-		name               string
-		setupCommands      func() []*cobra.Command
-		enabledFeatures    []string
-		expectedCmdNames   []string
-		expectedNotPresent []string
-	}{
-		{
-			name: "adds ungated commands",
-			setupCommands: func() []*cobra.Command {
-				return []*cobra.Command{
-					{Use: "ungated"},
-				}
-			},
-			expectedCmdNames: []string{"ungated"},
-		},
-		{
-			name: "filters out disabled gated commands",
-			setupCommands: func() []*cobra.Command {
-				gated := &cobra.Command{Use: "gated"}
-				SetGate(gated, "my-feature")
-
-				return []*cobra.Command{
-					{Use: "ungated"},
-					gated,
-				}
-			},
-			enabledFeatures:    []string{},
-			expectedCmdNames:   []string{"ungated"},
-			expectedNotPresent: []string{"gated"},
-		},
-		{
-			name: "adds enabled gated commands",
-			setupCommands: func() []*cobra.Command {
-				gated := &cobra.Command{Use: "gated"}
-				SetGate(gated, "my-feature")
-
-				return []*cobra.Command{
-					{Use: "ungated"},
-					gated,
-				}
-			},
-			enabledFeatures:  []string{"MY_FEATURE"},
-			expectedCmdNames: []string{"ungated", "gated"},
-		},
-		{
-			name: "handles nested gated commands with CommandAdder wrapper",
-			setupCommands: func() []*cobra.Command {
-				parent := &cobra.Command{Use: "parent"}
-
-				gatedChild := &cobra.Command{Use: "gated-child"}
-				SetGate(gatedChild, "child-feature")
-
-				parentAdder := &CommandAdder{Command: parent}
-				parentAdder.AddCommand(
-					&cobra.Command{Use: "ungated-child"},
-					gatedChild,
-				)
-
-				return []*cobra.Command{parent}
-			},
-			enabledFeatures:    []string{},
-			expectedCmdNames:   []string{"parent"},
-			expectedNotPresent: []string{"gated-child"},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Set env vars for enabled features
-			for _, feature := range tt.enabledFeatures {
-				t.Setenv("DATAROBOT_CLI_FEATURES_"+feature, "true")
-			}
-
-			root := &CommandAdder{
-				Command: &cobra.Command{Use: "root"},
-			}
-			root.AddCommand(tt.setupCommands()...)
-
-			// Check expected commands are present
-			for _, expectedName := range tt.expectedCmdNames {
-				found := false
-
-				for _, cmd := range root.Commands() {
-					if cmd.Name() == expectedName {
-						found = true
-						break
-					}
-				}
-
-				assert.True(t, found, "expected command %s to be present", expectedName)
-			}
-
-			// Check expected not present are absent
-			for _, notExpectedName := range tt.expectedNotPresent {
-				found := false
-
-				for _, cmd := range root.Commands() {
-					if cmd.Name() == notExpectedName {
-						found = true
-						break
-					}
-				}
-
-				assert.False(t, found, "expected command %s to be absent", notExpectedName)
-			}
-		})
-	}
-}
-
 func TestSetGate(t *testing.T) {
 	tests := []struct {
 		name             string
@@ -225,122 +114,6 @@ func TestSetGate(t *testing.T) {
 
 			if tt.shouldPreserveAn {
 				assert.Equal(t, "value", tt.cmd.Annotations["key"])
-			}
-		})
-	}
-}
-
-func TestRemoveDisabledCommands(t *testing.T) {
-	tests := []struct {
-		name                string
-		setupCmd            func() *cobra.Command
-		enabledFeatures     []string
-		expectedSubcommands []string
-		expectedNotPresent  []string
-	}{
-		{
-			name: "removes top-level gated command",
-			setupCmd: func() *cobra.Command {
-				parent := &cobra.Command{Use: "root"}
-
-				enabled := &cobra.Command{Use: "enabled"}
-				parent.AddCommand(enabled)
-
-				gated := &cobra.Command{
-					Use: "gated",
-					Annotations: map[string]string{
-						AnnotationKey: "test-feature",
-					},
-				}
-				parent.AddCommand(gated)
-
-				return parent
-			},
-			enabledFeatures:     []string{},
-			expectedSubcommands: []string{"enabled"},
-			expectedNotPresent:  []string{"gated"},
-		},
-		{
-			name: "keeps top-level gated command when enabled",
-			setupCmd: func() *cobra.Command {
-				parent := &cobra.Command{Use: "root"}
-
-				gated := &cobra.Command{
-					Use: "gated",
-					Annotations: map[string]string{
-						AnnotationKey: "test-feature",
-					},
-				}
-				parent.AddCommand(gated)
-
-				return parent
-			},
-			enabledFeatures:     []string{},
-			expectedSubcommands: []string{}, // Will be set in test
-			expectedNotPresent:  []string{},
-		},
-		{
-			name: "removes nested gated command",
-			setupCmd: func() *cobra.Command {
-				root := &cobra.Command{Use: "root"}
-
-				parent := &cobra.Command{Use: "parent"}
-				root.AddCommand(parent)
-
-				enabled := &cobra.Command{Use: "enabled"}
-				parent.AddCommand(enabled)
-
-				gated := &cobra.Command{
-					Use: "gated",
-					Annotations: map[string]string{
-						AnnotationKey: "nested-feature",
-					},
-				}
-				parent.AddCommand(gated)
-
-				return root
-			},
-			enabledFeatures:    []string{},
-			expectedNotPresent: []string{"gated"},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Set env vars for enabled features
-			for _, feature := range tt.enabledFeatures {
-				t.Setenv("DATAROBOT_CLI_FEATURES_"+feature, "true")
-			}
-
-			cmd := tt.setupCmd()
-			RemoveDisabledCommands(cmd)
-
-			// Check expected subcommands are present
-			for _, expectedName := range tt.expectedSubcommands {
-				found := false
-
-				for _, sub := range cmd.Commands() {
-					if sub.Name() == expectedName {
-						found = true
-						break
-					}
-				}
-
-				assert.True(t, found, "expected subcommand %s to be present", expectedName)
-			}
-
-			// Check expected not present are absent
-			for _, notExpectedName := range tt.expectedNotPresent {
-				found := false
-
-				for _, sub := range cmd.Commands() {
-					if sub.Name() == notExpectedName {
-						found = true
-						break
-					}
-				}
-
-				assert.False(t, found, "expected subcommand %s to be absent", notExpectedName)
 			}
 		})
 	}
